requirements.txt
flask
flask-cors
firebase-admin
requests
beautifulsoup4
waitress
gunicorn
python-dotenv
google-generativeai